fix: move event specific helper to event transformer
