--- app/site-config/event/event.helpers.ts
@@ -1,10 +1,0 @@
-import { makeCardMiniProps } from "../content.helpers";
-import type { EventContent } from "../types";
-
-export const transformEventToCardMiniProps = (event: EventContent) => {
-  const { isLatest, ...eventRest } = event;
-  return makeCardMiniProps({
-    ...eventRest,
-    ...(isLatest ? { tag: "Latest Activation" } : {}),
-  });
-};
